Snapshot and Record Upd

+++ Assets/Scripts/Storage/Records/LevelRecord.cs
@@ -5,7 +5,7 @@ namespace Storage.Records
     [Serializable]
     public class LevelRecord
     {
-        public LevelRecord(int levelNumber, float bestCompletedTime, int starsEarned, UnlockStatus isUnlocked, int playCount)
+        public LevelRecord(int levelNumber, double bestCompletedTime, int starsEarned, UnlockStatus isUnlocked, int playCount)
         {
             LevelNumber = levelNumber;
             BestCompletedTime = bestCompletedTime;
@@ -16,7 +16,7 @@ namespace Storage.Records
         
         public int LevelNumber;
         public int StarsEarned;
-        public float BestCompletedTime;
+        public double BestCompletedTime;
         public int PlayCount;
         
         //TODO IN THE FUTURE BLOCK
